+++ backend/src/main/java/ca/bc/gov/nrs/ilcr/schedule1/dto/Schedule1Response.java
@@ -28,7 +28,7 @@ public record Schedule1Response(
     String comments,
     List<LineItem> lineItems,
     SilvicultureBlock silviculture,
-    Integer forestMgmtAdminCost,
+    Long forestMgmtAdminCost,
     Integer lessSilvAdminCost,
     OtherCostsSummary otherCosts,
     // Derived read-only figures (legacy Schedule1MB getters) — the $/m³ ("Cal") per-unit cells and the
